# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WiseEff API",
    "version": "m5"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "parameters.listProjects",
        "summary": "List projects",
        "tags": [
          "parameters"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/modules": {
      "get": {
        "operationId": "parameters.listProjectModules",
        "summary": "List project modules",
        "tags": [
          "parameters"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModuleListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameters": {
      "get": {
        "operationId": "parameters.list",
        "summary": "List parameters",
        "tags": [
          "parameters"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameters/{parameterId}": {
      "get": {
        "operationId": "parameters.get",
        "summary": "Get parameter",
        "tags": [
          "parameters"
        ],
        "parameters": [
          {
            "name": "parameterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameters/{parameterId}/history": {
      "get": {
        "operationId": "parameters.history",
        "summary": "Get parameter history",
        "tags": [
          "parameters"
        ],
        "parameters": [
          {
            "name": "parameterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterHistoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-drafts": {
      "post": {
        "operationId": "parameters.saveDraft",
        "summary": "Save parameter draft",
        "tags": [
          "parameters"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveParameterDraftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterDraftResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-drafts/mine": {
      "get": {
        "operationId": "parameters.listMyDrafts",
        "summary": "List my parameter drafts",
        "tags": [
          "parameters"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterDraftListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-drafts/{draftId}": {
      "delete": {
        "operationId": "parameters.deleteDraft",
        "summary": "Delete parameter draft",
        "tags": [
          "parameters"
        ],
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-submission-rounds": {
      "post": {
        "operationId": "parameters.submitRound",
        "summary": "Submit parameter review round",
        "tags": [
          "parameters"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitParameterRoundRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterSubmissionRoundResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "get": {
        "operationId": "parameters.listSubmissionRounds",
        "summary": "List parameter submission rounds",
        "tags": [
          "parameters"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterSubmissionRoundListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-change-requests": {
      "get": {
        "operationId": "parameters.listChangeRequests",
        "summary": "List parameter change requests",
        "tags": [
          "parameters"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterChangeRequestListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-change-requests/{requestId}/review": {
      "post": {
        "operationId": "parameters.reviewChangeRequest",
        "summary": "Review parameter change request",
        "tags": [
          "parameters"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewParameterChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterChangeRequestResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-import-batches": {
      "post": {
        "operationId": "parameters.createImportBatch",
        "summary": "Create parameter import batch",
        "tags": [
          "parameters"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParameterImportBatchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterImportBatchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/parameter-import-batches/{batchId}/apply": {
      "post": {
        "operationId": "parameters.applyImportBatch",
        "summary": "Apply parameter import batch",
        "tags": [
          "parameters"
        ],
        "parameters": [
          {
            "name": "batchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyParameterImportBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterImportBatchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/debugging/parameters": {
      "get": {
        "operationId": "debugging.listParameters",
        "summary": "List debug parameters",
        "tags": [
          "debugging"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugParameterListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProjectListResponse": {
        "type": "object",
        "x-wiseeff-schema": "ProjectListResponse"
      },
      "ProjectModuleListResponse": {
        "type": "object",
        "x-wiseeff-schema": "ProjectModuleListResponse"
      },
      "ParameterListResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterListResponse"
      },
      "ParameterResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterResponse"
      },
      "ParameterHistoryResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterHistoryResponse"
      },
      "SaveParameterDraftRequest": {
        "type": "object",
        "x-wiseeff-schema": "SaveParameterDraftRequest"
      },
      "ParameterDraftResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterDraftResponse"
      },
      "ParameterDraftListResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterDraftListResponse"
      },
      "DeleteResponse": {
        "type": "object",
        "x-wiseeff-schema": "DeleteResponse"
      },
      "SubmitParameterRoundRequest": {
        "type": "object",
        "x-wiseeff-schema": "SubmitParameterRoundRequest"
      },
      "ParameterSubmissionRoundResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterSubmissionRoundResponse"
      },
      "ParameterSubmissionRoundListResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterSubmissionRoundListResponse"
      },
      "ParameterChangeRequestListResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterChangeRequestListResponse"
      },
      "ReviewParameterChangeRequest": {
        "type": "object",
        "x-wiseeff-schema": "ReviewParameterChangeRequest"
      },
      "ParameterChangeRequestResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterChangeRequestResponse"
      },
      "CreateParameterImportBatchRequest": {
        "type": "object",
        "x-wiseeff-schema": "CreateParameterImportBatchRequest"
      },
      "ParameterImportBatchResponse": {
        "type": "object",
        "x-wiseeff-schema": "ParameterImportBatchResponse"
      },
      "ApplyParameterImportBatchRequest": {
        "type": "object",
        "x-wiseeff-schema": "ApplyParameterImportBatchRequest"
      },
...
      "DebugParameterListResponse": {
        "type": "object",
        "x-wiseeff-schema": "DebugParameterListResponse"
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "WiseEff error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}